requirements.txt
streamlit
streamlit-option-menu
google-api-python-client
mysql-connector-python
pandas
python-dotenv
isodate

